app/backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.4
requests==2.31.0
httpx==0.27.0
django-cors-headers==4.3.1
django-filter==24.2
gunicorn==21.2.0
pytest==8.1.1
pytest-django==4.8.0
pre-commit==3.7.0
black==24.4.2
flake8==7.0.0
isort==5.13.2
